Fix ipopt cmake
This "should" solve the cmake problem and fix feel++ issue
https://github.com/feelpp/feelpp/issues/956

Note:

 This custom ipopt CMakeLists.txt parse all Makefile.am
files (ipopt is autoconf based) line by line to retrieve ipopt
sources and headers.

Headers are first generated (creating a copy) then used.
The ipopt configuration file (another header file) is generated
by moving one of this previous generated header.

The problem seems to be that these two operations are realized
line by line at the same time meaning that the header does not
yet exist, when the configuration is generated.
(which explain why 2 cmake was required to make it works).
The problem might have appears after an update if two sources
of the Makefile.am have changed order.